[seller] refactor(230): 서류 등록 공통 onError 토스트를 mutation hook으로 이동
- 페이지 state에 의존하지 않는 에러 처리를 mutation hook 기본값으로 옮기고 페이지는 페이지 고유 onSuccess만 담당하도록 정리
